requirements.txt
langchain
langchain-community
pymupdf
sentence-transformers 
faiss-cpu 
groq
langchain-groq
python-dotenv
streamlit